pyproject.toml
[tool.poetry]
name = "roses"
version = "0.1.0"
description = ""
authors = ["bezro <bezrogota@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "==3.10"
scikit-learn = "1.3.0"
pre-commit = "2.16.0"
dvc = {extras = ["gdrive"], version = "^3.25.0"}
hydra-core = "^1.3.2"
fire = "^0.5.0"
mlflow = "^2.8.1"
torch = "^2.1.1"
onnx = "^1.15.0"
onnxruntime = "^1.16.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
